+++ deck_chores/exceptions.py
@@ -1,2 +1,6 @@
 class ConfigurationError(Exception):
     pass
+
+
+class ParsingError(Exception):
+    pass