+++ deck_chores/indexes.py
@@ -1,0 +1,2 @@
+locking_container_to_services_map = {}  # type: Dict[str, str]
+""" A mapping of locking container ids to a service ids whose jobs have been added. """