+++ HISTORY.rst
@@ -1,6 +1,11 @@
 History
 =======
 
+unreleased
+----------
+
+* *new:* set log format per :envvar:LOG_FORMAT
+
 0.1.beta1 (2016-12-04)
 ----------------------