+++ HISTORY.rst
@@ -5,6 +5,7 @@ unreleased
 ----------
 
 * *new:* set log format per :envvar:LOG_FORMAT
+* *new:* an options label to set behavioural flags
 
 0.1.beta1 (2016-12-04)
 ----------------------