+++ HISTORY.rst
@@ -7,6 +7,8 @@ unreleased
 * *new:* set log format per :envvar:LOG_FORMAT
 * *new:* an options label to set behavioural flags
 * *new:* containers can be identified as a service by configurable labels
+* *new:* job definitions for further containers of a service are ignored
+         (default, opt-out can be configured)
 
 0.1.beta1 (2016-12-04)
 ----------------------