+++ HISTORY.rst
@@ -9,6 +9,8 @@ unreleased
 * *new:* containers can be identified as a service by configurable labels
 * *new:* job definitions for further containers of a service are ignored
          (default, opt-out can be configured)
+* *new:* image labels can also be parsed for job definitions
+         (default, opt-out can be configured)
 
 0.1.beta1 (2016-12-04)
 ----------------------